move json parsing code to own src file
first step of refactoring processing moving the very long json schema
parsing function to its own source file
